add specification module
