+++ src/fmu/dataio/models/meta/specification.py
@@ -1,0 +1,155 @@
+from __future__ import annotations
+
+from typing import Union
+
+from pydantic import BaseModel, Field
+
+from . import enums
+
+
+class Shape(BaseModel):
+    nrow: int = Field(
+        description="The number of rows",
+    )
+    ncol: int = Field(
+        description="The number of columns",
+    )
+    nlay: int = Field(
+        description="The number of layers",
+    )
+
+
+class SurfaceSpecification(Shape):
+    rotation: float = Field(
+        description="Rotation angle",
+        allow_inf_nan=False,
+    )
+    undef: float = Field(
+        description="Value representing undefined data",
+        allow_inf_nan=False,
+    )
+    xinc: float = Field(
+        description="Increment along the x-axis",
+        allow_inf_nan=False,
+    )
+    xori: float = Field(
+        description="Origin along the x-axis",
+        allow_inf_nan=False,
+    )
+    yflip: enums.AxisOrientation = Field(
+        description="Flip along the y-axis, -1 or 1",
+    )
+    yori: float = Field(
+        description="Origin along the y-axis",
+        allow_inf_nan=False,
+    )
+
+
+class TableSpecification(BaseModel):
+    columns: list[str] = Field(
+        description="List of columns present in a table.",
+    )
+    size: int = Field(
+        description="Size of data object.",
+        examples=[1, 9999],
+    )
+
+
+class CPGridSpecification(Shape):
+    """Corner point grid"""
+
+    xshift: float = Field(
+        description="Shift along the x-axis",
+        allow_inf_nan=False,
+    )
+    yshift: float = Field(
+        description="Shift along the y-axis",
+        allow_inf_nan=False,
+    )
+    zshift: float = Field(
+        description="Shift along the z-axis",
+        allow_inf_nan=False,
+    )
+
+    xscale: float = Field(
+        description="Scaling factor for the x-axis",
+        allow_inf_nan=False,
+    )
+    yscale: float = Field(
+        description="Scaling factor for the y-axis",
+        allow_inf_nan=False,
+    )
+    zscale: float = Field(
+        description="Scaling factor for the z-axis",
+        allow_inf_nan=False,
+    )
+
+
+class CPGridPropertySpecification(Shape):
+    ...
+
+
+class PolygonsSpecification(BaseModel):
+    npolys: int = Field(
+        description="The number of individual polygons in the data object",
+    )
+
+
+class CubeSpecification(SurfaceSpecification):
+    # Increment
+    xinc: float = Field(
+        description="Increment along the x-axis",
+        allow_inf_nan=False,
+    )
+    yinc: float = Field(
+        description="Increment along the y-axis",
+        allow_inf_nan=False,
+    )
+    zinc: float = Field(
+        description="Increment along the z-axis",
+        allow_inf_nan=False,
+    )
+
+    # Origin
+    xori: float = Field(
+        description="Origin along the x-axis",
+        allow_inf_nan=False,
+    )
+    yori: float = Field(
+        description="Origin along the y-axis",
+        allow_inf_nan=False,
+    )
+    zori: float = Field(
+        description="Origin along the z-axis",
+        allow_inf_nan=False,
+    )
+
+    # Miscellaneous
+    yflip: enums.AxisOrientation = Field(
+        description="Flip along the y-axis, -1 or 1",
+    )
+    zflip: enums.AxisOrientation = Field(
+        description="Flip along the z-axis, -1 or 1",
+    )
+    rotation: float = Field(
+        description="Rotation angle",
+        allow_inf_nan=False,
+    )
+    undef: float = Field(
+        description="Value representing undefined data",
+    )
+
+
+class WellPointsDictionaryCaseSpecification(BaseModel):
+    ...
+
+
+AnySpecification = Union[
+    CPGridPropertySpecification,
+    CPGridSpecification,
+    CubeSpecification,
+    PolygonsSpecification,
+    SurfaceSpecification,
+    TableSpecification,
+    WellPointsDictionaryCaseSpecification,
+]
